+++ src/App.jsx
@@ -1,8 +1,8 @@
+import Input from './components/input';
 function App() {
   return (
     <>
-      {/* <Input placeholder="플레이스호더" errorMassage="에러다" id="1"></Input>
-      <DropDown label="친구" name="relationship" placeholders={a} id="1" /> */}
+      <Input placeholder="플레이스호더" errorMassage="에러다"></Input>
     </>
   );
 }
